third_party/WebKit/LayoutTests/accessibility/not-ignore-landmark-roles-expected.txt

article element
section element
header element
aside element
address element
footer element
main element
nav element
This tests that landmark and related roles are not ignored.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS containerNode.childrenCount is 8
PASS AXObjectImpl.role is 'AXRole: AXArticle'
PASS AXObjectImpl.role is 'AXRole: AXRegion'
PASS AXObjectImpl.role is 'AXRole: AXBanner'
PASS AXObjectImpl.role is 'AXRole: AXComplementary'
PASS AXObjectImpl.role is 'AXRole: AXContentInfo'
PASS AXObjectImpl.role is 'AXRole: AXFooter'
PASS AXObjectImpl.role is 'AXRole: AXMain'
PASS AXObjectImpl.role is 'AXRole: AXNavigation'
PASS successfullyParsed is true

TEST COMPLETE
